# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Push API",
      "version": "1.0.0",
      "description": "API for sending push notifications to iOS, Android, and web applications"
    },
    "servers": [
      {
        "url": "https://api.push.com/v1",
        "description": "Production server"
      }
    ],
    "paths": {
...
      "/messages": {
        "post": {
          "summary": "Send Message",
          "description": "Send a push notification to one or more devices",
          "operationId": "sendMessage",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["userId", "title", "body"],
                  "properties": {
                    "userId": {
                      "type": "string",
                      "description": "User ID to send the notification to"
                    },
                    "title": {
                      "type": "string",
                      "description": "Title of the notification"
                    },
                    "body": {
                      "type": "string",
                      "description": "Body of the notification"
                    },
                    "platform": {
                      "type": "string",
                      "enum": ["ios", "android", "web"],
                      "description": "Platform to send the notification to (optional, sends to all platforms if not specified)"
                    },
                    "deviceIdentifier": {
                      "type": "string",
                      "description": "Specific device to send the notification to (optional, sends to all user's devices if not specified)"
                    },
                    "category": {
                      "type": "string",
                      "description": "Category of the notification (e.g., 'marketing', 'security', 'transaction') (optional)"
                    },
                    "data": {
                      "type": "object",
                      "description": "Additional data to send with the notification"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Message sent successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Unique identifier for the sent message"
                      },
                      "userId": {
                        "type": "string",
                        "description": "User ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "Message title"
                      },
                      "body": {
                        "type": "string",
                        "description": "Message body"
                      },
                      "platform": {
                        "type": "string",
                        "enum": ["ios", "android", "web"],
                        "description": "Platform"
                      },
                      "deviceIdentifier": {
                        "type": "string",
                        "description": "Device identifier"
                      },
                      "category": {
                        "type": "string",
                        "description": "Category of the notification"
                      },
                      "data": {
                        "type": "object",
                        "description": "Additional data sent with the notification"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Sent date"
                      },
                      "devices": {
                        "type": "array",
                        "description": "Devices the message was sent to",
                        "items": {
                          "type": "object",
                          "properties": {
                            "deviceIdentifier": {
                              "type": "string",
                              "description": "Device identifier"
                            },
                            "platform": {
                              "type": "string",
                              "enum": ["ios", "android", "web"],
                              "description": "Platform"
                            },
                            "status": {
                              "type": "string",
                              "enum": ["sent", "delivered", "failed"],
                              "description": "Delivery status"
                            },
                            "deliveredAt": {
                              "type": "string",
                              "format": "date-time",
                              "description": "Delivery date"
                            },
                            "error": {
                              "type": "string",
                              "description": "Error message if delivery failed"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request parameters"
            },
            "401": {
              "description": "Unauthorized - Invalid API key"
            }
          }
        },
        "get": {
          "summary": "List Messages",
          "description": "Get a list of sent messages with optional filtering",
          "operationId": "listMessages",
          "parameters": [
            {
              "name": "userId",
              "in": "query",
              "description": "User ID to list messages for (optional)",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "platform",
              "in": "query",
              "description": "Filter messages by platform (optional)",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["ios", "android", "web"]
              }
            },
            {
              "name": "deviceIdentifier",
              "in": "query",
              "description": "Filter messages by device (optional)",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "category",
              "in": "query",
              "description": "Filter messages by category (optional)",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "status",
              "in": "query",
              "description": "Filter messages by delivery status (optional)",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["sent", "delivered", "failed"]
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Maximum number of messages to return",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 10
              }
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Number of messages to skip",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 0
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of messages",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "Message ID"
                            },
                            "userId": {
                              "type": "string",
                              "description": "User ID"
                            },
                            "title": {
                              "type": "string",
                              "description": "Message title"
                            },
                            "body": {
                              "type": "string",
                              "description": "Message body"
                            },
                            "platform": {
                              "type": "string",
                              "enum": ["ios", "android", "web"],
                              "description": "Platform"
                            },
                            "deviceIdentifier": {
                              "type": "string",
                              "description": "Device identifier"
                            },
                            "category": {
                              "type": "string",
                              "description": "Category of the notification"
                            },
                            "data": {
                              "type": "object",
                              "description": "Additional data sent with the notification"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time",
                              "description": "Sent date"
                            },
                            "devices": {
                              "type": "array",
                              "description": "Devices the message was sent to",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "deviceIdentifier": {
                                    "type": "string",
                                    "description": "Device identifier"
                                  },
                                  "platform": {
                                    "type": "string",
                                    "enum": ["ios", "android", "web"],
                                    "description": "Platform"
                                  },
                                  "status": {
                                    "type": "string",
                                    "enum": ["sent", "delivered", "failed"],
                                    "description": "Delivery status"
                                  },
                                  "deliveredAt": {
                                    "type": "string",
                                    "format": "date-time",
                                    "description": "Delivery date"
                                  },
                                  "error": {
                                    "type": "string",
                                    "description": "Error message if delivery failed"
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "total": {
                        "type": "integer",
                        "description": "Total number of messages"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized - Invalid API key"
            }
          }
        }
      },
      "/messages/{id}": {
        "get": {
          "summary": "Get Message",
          "description": "Get details about a specific message",
          "operationId": "getMessage",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Message ID",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Message details",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Message ID"
                      },
                      "userId": {
                        "type": "string",
                        "description": "User ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "Message title"
                      },
                      "body": {
                        "type": "string",
                        "description": "Message body"
                      },
                      "platform": {
                        "type": "string",
                        "enum": ["ios", "android", "web"],
                        "description": "Platform"
                      },
                      "deviceIdentifier": {
                        "type": "string",
                        "description": "Device identifier"
                      },
                      "category": {
                        "type": "string",
                        "description": "Category of the notification"
                      },
                      "data": {
                        "type": "object",
                        "description": "Additional data sent with the notification"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Sent date"
                      },
                      "devices": {
                        "type": "array",
                        "description": "Devices the message was sent to",
                        "items": {
                          "type": "object",
                          "properties": {
                            "deviceIdentifier": {
                              "type": "string",
                              "description": "Device identifier"
                            },
                            "platform": {
                              "type": "string",
                              "enum": ["ios", "android", "web"],
                              "description": "Platform"
                            },
                            "status": {
                              "type": "string",
                              "enum": ["sent", "delivered", "failed"],
                              "description": "Delivery status"
                            },
                            "deliveredAt": {
                              "type": "string",
                              "format": "date-time",
                              "description": "Delivery date"
                            },
                            "error": {
                              "type": "string",
                              "description": "Error message if delivery failed"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Message not found"
            },
            "401": {
              "description": "Unauthorized - Invalid API key"
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "apiKey": {
          "type": "apiKey",
          "in": "header",
          "name": "Authorization",
          "description": "API key for authentication"
        }
      }
    },
    "security": [
      {
        "apiKey": []
      }
    ]
  }